Remove aset and sset in lieu of proper lvalues/rvalues
Working on the MARS backend made me realize that aset and sset are
unnecessary, since the part of the backend which knows how to deal with
expressions also has to know how to deal with assignment targets. Since the
backend can easily differentiate between 'assignable expressions' and
'non-assignable expressions' (C's lvalues and rvalues), keeping around
separate assignment constructs is unnecessary.