Remove function/value distinction
    (declare (g (function byte))
             (f (function byte))
    (define f ()
     (declare g (func-ptr byte))
     (begin
      (g)))

This fragment is ambiguous (does g call the toplevle function or the
pointer?), and there are two ways to fix it - either require all functions
and values to be unique between each other, or have a special CL-like
funcall form.

Neither of these options is particularly attractive; instead, functions and
values will live in the same namespace, and functions will automatically
be converted to function pointers when used.