Add special handling for loading array values
Array values are weird when in their declared scope - they are loaded
not like normal pointers (which are integers in memory), but more like
references to values (which are computed memory locations). Thus, they
should always be by_ref

However, since arrays decay immediately, the only way to know about which
variables are arrays and which are not is by stuffing them into the current
context so they can be looked up later.