Found a parameter reversing bug via the multi-param sample
This one is kind of subtle - the FunctionStack takes care of parameter
reversal on its own (since the parameter list starts at 0 and grows up)
which mirrors the reversal in parameter passing. Doing a double reversal
means that parameters are assigned locations that are non-reversed even
though they should be.