Generalize 32-bit backend
To solve the high amount of code duplication that came with the
linux_x86 backend, all the places where the mars backend emitted code
were replaced with calls to a template object which actually emits code.

This means that the logic for emitting 32-bit code can be shared across
both mars and linux_x86, by  creating a new template object which emits
the proper code for each method.

Currently, only the mars backend has been moved over to the template
method. It isn't crashing, but it is generating invalid code in some
circumstances.