Add *if and *error
*if, or static if, allows for conditional compilation depending upon
conditions which are resolvable at compile time.

The motivation is for allowing the development of platform-independent
code, with the weakest possible mechanism. Rather than embedding a
complete macro processor (like C's preprocessor), *if allows only for
conditional compilation:

    (*if (platform? "linux_x86")
     ;; Do Linux things
     ;; Do non-Linux things)

This also includes *error, for raising errors at compile time. This
makes it possible to have explicit errors, in case necessary *if
conditions fail:

    (*if (platform? "linux_x86")
     ;; Do Linux things
     (*error "This module supported only on Linux"))