Add shallow symbol table iteration
This stems out of a need to handle builtin types differently - because
of the way that 'string' is defined (not as a builtin, but by injecting it
into the base symbol table), it has to be added specially to the require
processor to be available during the 'require' phase.

However, this caused an issue where 'string' had to be handled
specially, since it would effectively be exported by every module,
causing symbol clashes.

Thus, 'string' is put in a "base" symbol table (above the export list),
and symbol tables grew a 'shallow iterator' so that iterating over the
export list would not get 'string' from the base symbol table above it.