Make visibility more lenient
Previously, visibility was restricted to the current namespace only -
now, the visibility checker considers parent namespaces as well. This
allows the backend to get away with fewer visibility annotations. For
example:

    (declare (x integer)
             (foo (function ...)))

    (define foo ()
     ...)

If foo wanted to see x (as it should), the backend would have had to add
foo's namespace to x's visibility list. Now, since foo's context has the
file's namespace in its search list, x is visible without any extra work
on the backend's part (other than making x visible to the file when it
is declared)