Add a main stub
With the introduction of namespaces, the name for main is not just 'main',
but the namespace and then 'main' within that namespace. So, instead of
having a simple main function whose name is known in advance (allowing the
backend to output it at the beginning of the assembly file), there is an
additional main stub which can be output at the end of the assembly file
once the full name of 'main' is known.